feat: add rate limit to api routes

+++ src/app/api/post/[id]/route.ts
@@ -1,14 +1,26 @@
 import { prisma } from '@libs/prisma'
 import { getSession } from '@libs/auth'
 import { NextResponse } from 'next/server'
+import { rateLimit } from '@libs/rate-limit'
 import { CreatePostSchema } from '@validations/create-post-schema'
 import { deleteFromCloudinary, uploadToCloudinary } from '@libs/cloudinary'
+
+const limiter = rateLimit({
+	limit: 5,
+	interval: 60 * 60 * 1000 // 1 hour
+})
 
 export async function PATCH(request: Request, { params }: { params: { id: string } }) {
 	const session = await getSession()
 
 	if (!session) {
 		return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
+	}
+
+	const { isRateLimited } = limiter.check(session.user.id)
+
+	if (isRateLimited) {
+		return NextResponse.json({ error: 'Too many requests' }, { status: 429 })
 	}
 
 	if (request.headers.get('content-type')?.includes('multipart/form-data')) {
@@ -65,6 +77,12 @@ export async function POST(request: Request, { params }: { params: { id: string 
 		return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
 	}
 
+	const { isRateLimited } = limiter.check(session.user.id)
+
+	if (isRateLimited) {
+		return NextResponse.json({ error: 'Too many requests' }, { status: 429 })
+	}
+
 	const postsCount = await prisma.post.count({
 		where: {
 			siteId: params.id
@@ -118,6 +136,12 @@ export async function DELETE(request: Request, { params }: { params: { id: strin
 
 	if (!session) {
 		return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
+	}
+
+	const { isRateLimited } = limiter.check(session.user.id)
+
+	if (isRateLimited) {
+		return NextResponse.json({ error: 'Too many requests' }, { status: 429 })
 	}
 
 	const post = await prisma.post.delete({
